+++ core/src/main/kotlin/com/toasttab/expediter/provider/ApplicationTypesProvider.kt
@@ -13,14 +13,10 @@
  * limitations under the License.
  */
 
-package com.toasttab.expediter.types
+package com.toasttab.expediter.provider
 
-import protokt.v1.toasttab.expediter.v1.TypeDescriptor
+import com.toasttab.expediter.types.ApplicationType
 
-class InMemoryPlatformTypeProvider private constructor(
-    private val types: Map<String, TypeDescriptor>
-) : PlatformTypeProvider {
-    constructor(types: Collection<TypeDescriptor>) : this(types.associateBy { it.name })
-
-    override fun lookupPlatformType(name: String) = types[name]
+interface ApplicationTypesProvider {
+    fun types(): List<ApplicationType>
 }
